Fix session-del!
Patch by Evan Hanson:

session-del! currently fails to remove an entry for the given key when
that entry is the first item in the pertinent session-item's bindings
alist (because alist-delete! just returns the alist's cdr when its car
is to be removed (without set-cdr!ing anything) and the session-item's
bindings slot is never updated to refer to the new start of the alist).

This patch fixes #994 (http://bugs.call-cc.org/ticket/994)